added loading elements on feed, upload profile fix

+++ app/src/main/java/sk/stuba/fei/mv/android/zaverecne/feed/FeedFragment.kt
@@ -2,16 +2,18 @@ package sk.stuba.fei.mv.android.zaverecne.feed
 
 import android.Manifest
 import android.app.Activity
-import android.app.ActivityOptions
+import android.content.Context
+import android.content.DialogInterface
 import android.content.Intent
 import android.content.pm.PackageManager
-import android.graphics.Bitmap
 import android.os.Bundle
 import android.util.Log
 import android.view.*
 import android.view.animation.AnimationUtils
 import android.view.animation.LayoutAnimationController
 import android.widget.LinearLayout
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
@@ -22,13 +24,12 @@ import com.leinardi.android.speeddial.SpeedDialActionItem
 import com.leinardi.android.speeddial.SpeedDialView.OnActionSelectedListener
 import kotlinx.android.synthetic.main.feed_fragment.*
 import kotlinx.android.synthetic.main.profile_fragment.*
+import kotlinx.android.synthetic.main.video_fragment.*
+import sk.stuba.fei.mv.android.zaverecne.MainActivity
 import sk.stuba.fei.mv.android.zaverecne.R
 import sk.stuba.fei.mv.android.zaverecne.databinding.FeedFragmentBinding
-import sk.stuba.fei.mv.android.zaverecne.fetchfiles.FetchFiles.getPath
 import sk.stuba.fei.mv.android.zaverecne.fetchfiles.FetchFiles.getRealPathFromURI
-import sk.stuba.fei.mv.android.zaverecne.gallery.FolderRecycleView
-import sk.stuba.fei.mv.android.zaverecne.profile.ProfileFragment
-import java.io.File
+import sk.stuba.fei.mv.android.zaverecne.repository.MasterRepository
 
 
 class FeedFragment : Fragment(){
@@ -105,7 +106,7 @@ class FeedFragment : Fragment(){
 
     fun showMenu(feedViewModel: FeedViewModel, feedPost: FeedPost) {
         val dialog = RoundedBottomSheetDialog(context!!)
-        val btnsheet = layoutInflater.inflate(R.layout.bottom_sheet_dialog_post, null)
+        val btnsheet = layoutInflater.inflate(R.layout.bottom_sheet_dialog_fragment, null)
         dialog.setContentView(btnsheet)
         btnsheet.setOnClickListener {
         }
@@ -114,9 +115,57 @@ class FeedFragment : Fragment(){
         val deletePostBtn =
             btnsheet.findViewById<LinearLayout>(R.id.deletePostButton)
         deletePostBtn.setOnClickListener(View.OnClickListener {
-            feedViewModel.deleteUserPost(feed, feedPost)
+            deleteVideoDialog(feedViewModel, feedPost)
             dialog.dismiss()
         })
+
+        val sharePostBtn =
+                btnsheet.findViewById<LinearLayout>(R.id.shareButtonPost)
+        sharePostBtn.setOnClickListener(View.OnClickListener {
+            val url = MasterRepository.mediaUrlBase + feedPost.videoSrc
+            shareVideo(url)
+            dialog.dismiss()
+        })
+    }
+
+    private fun deleteVideoDialog(feedViewModel: FeedViewModel, feedPost: FeedPost) {
+        val alert: AlertDialog.Builder = AlertDialog.Builder(context!!)
+        val linearLayout = LinearLayout(context)
+        linearLayout.orientation = LinearLayout.VERTICAL
+        val lp = LinearLayout.LayoutParams(LinearLayout.LayoutParams.MATCH_PARENT, LinearLayout.LayoutParams.WRAP_CONTENT)
+        lp.setMargins(50, 0, 100, 0)
+        alert.setMessage("Tento príspevok bude odstránený a už ho nebudete môcť nájsť..")
+        alert.setView(linearLayout)
+
+        alert.setNegativeButton("Zrušiť", DialogInterface.OnClickListener { dialog, which -> dialog.dismiss() })
+        alert.setPositiveButton("Odstrániť", DialogInterface.OnClickListener { dialog, which ->
+                    feedViewModel.deleteUserPost(feed, feedPost)
+                    Toast.makeText(context, "The post has been deleted.", Toast.LENGTH_LONG).show()
+                    dialog.dismiss()
+                    feedViewModel.getUserPosts()
+        })
+        alert.show()
+
+    }
+
+    fun shareVideo(url: String?) {
+        //2
+        val intent = Intent(Intent.ACTION_SEND).apply {
+            //3
+            type = "video/mp4"
+            //4
+            putExtra(Intent.EXTRA_TEXT, url)
+            //5
+//            val uri = Uri.parse(url)
+//            putExtra(Intent.EXTRA_STREAM, uri)
+//            //6
+//            val videoUrl = URL(url)
+//            clipData = ClipData.newUri(context.contentResolver, context.getString(R.string.app_name), FileProvider.getUriForFile(context, FILE_PROVIDER, videoUrl))
+            //7
+            flags = Intent.FLAG_GRANT_READ_URI_PERMISSION
+        }
+        //8
+        context?.startActivity(Intent.createChooser(intent, null))
     }
 
     private fun captureVideo() {
@@ -135,8 +184,8 @@ class FeedFragment : Fragment(){
 
     private fun checkPermissionStorage(): Boolean {
         val result = ContextCompat.checkSelfPermission(
-            context!!,
-            Manifest.permission.READ_EXTERNAL_STORAGE
+                context!!,
+                Manifest.permission.READ_EXTERNAL_STORAGE
         )
         return result == PackageManager.PERMISSION_GRANTED
     }
@@ -149,28 +198,49 @@ class FeedFragment : Fragment(){
 
     private fun requestPermission() {
         ActivityCompat.requestPermissions(
-            activity!!,
-            arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE),
-            PERMISSION_REQUEST_CODE
+                activity!!,
+                arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE),
+                PERMISSION_REQUEST_CODE
         )
+    }
+
+    override fun onPause() {
+        super.onPause()
+        Log.d("state", "onPause")
+    }
+
+    override fun onResume() {
+        super.onResume()
+        Log.d("state", "onResume")
+    }
+
+    override fun onStop() {
+        super.onStop()
+        Log.d("state", "onStop")
+        releasePlayer()
+    }
+
+    private fun releasePlayer(){
+
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
         if (resultCode == Activity.RESULT_OK && requestCode == REQUEST_CODE){
             val path = data?.data
-            val realPath = getPath(activity,path)
+            val realPath = getRealPathFromURI(context!!, path!!)
+            Log.d("realPath", realPath.toString())
             val args = Bundle()
             args.putString("videoUri", realPath);
             val navController = findNavController();
-            navController.navigate(R.id.action_feedFragment_to_videoFragment,args)
+            navController.navigate(R.id.action_feedFragment_to_videoFragment, args)
         }
     }
 
     override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String?>,
-        grantResults: IntArray
+            requestCode: Int,
+            permissions: Array<String?>,
+            grantResults: IntArray
     ) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         when (requestCode) {
